requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.2
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
